pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "porki"
version = "0.8.0"
description = "Agentic orchestration."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "redis>=5.0",
    "fakeredis>=2.22",
    "pydantic>=2.8",
    "pyyaml>=6.0",
    "tiktoken>=0.7",
]

[project.scripts]
porki = "main:run_cli"

[tool.setuptools]
py-modules = ["main"]

[tool.setuptools.packages.find]
include = ["porki"]

[dependency-groups]
dev = [
    "pytest>=9.0",
    "ruff>=0.13",
    "build>=1.4.0",
    "twine>=6.2.0",
]

[tool.uv]
default-groups = ["dev"]

[tool.pytest.ini_options]
pythonpath = ["."]
filterwarnings = [
  "error::DeprecationWarning",
  "ignore:Call to '__init__' function with deprecated usage of input argument/s '.*':DeprecationWarning",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
]
ignore = [
    "E501", # line-too-long (handled by formatter)
]
